pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "deathbed"
version = "3.1.0"
description = "Find the files in your codebase that are dying before they kill you."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
authors = [{ name = "Nikolozi Khachiashvili" }]
keywords = ["git", "code-quality", "analysis", "cli", "developer-tools", "technical-debt"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Version Control :: Git",
]
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "gitpython>=3.1",
    "radon>=6.0",
    "pathspec>=0.11",
    "vulture>=2.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.4",
    "mypy>=1.0",
]
interactive = ["textual>=0.40"]

[project.scripts]
deathbed = "deathbed.cli:main"

[project.urls]
Homepage = "https://github.com/NikoloziKhachiashvili/deathbed"
Repository = "https://github.com/NikoloziKhachiashvili/deathbed"
Issues = "https://github.com/NikoloziKhachiashvili/deathbed/issues"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=deathbed --cov-report=term-missing -q"

[tool.coverage.run]
source = ["deathbed"]
omit = ["tests/*"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["E501"]  # long HTML/string lines are intentional in templates

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
